pyproject.toml
[project]
name = "test-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.2",
    "django>=6.0.2",
    "django-environ>=0.12.0",
    "django-import-export>=4.4.0",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "gunicorn>=25.0.3",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "redis>=7.1.0",
]
[tool.ruff]
line-length = 88
exclude = [
    "config/",
    "migrations/",
    "__pycache__/",
    ".git/",
    ".venv/",
    "venv/",
]

[tool.ruff.lint]
select = ["I", "C", "E", "F", "W"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py"]
